Create README.md
This is an R implementation of power posterior methods, such as thermodynamic integration and stepping-stone sampling, for estimating the marginal likelihood and subsequently model selection. The uncertainty associated to the estimates is estimated via moving block bootstrap.  Examples of the methodology can be found in Maturana, Meyer, Veitch and Christensen <arXiv:1810.04488>.